pyproject.toml
[tool.poetry]
name = "python-flask"
version = "0.1.0"
description = "Start a simple web app with the Flask framework for Python"
authors = ["Google LLC"]
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
Flask = "^3.0.1"
requests = "^2.32.4"
sqlalchemy = "^2.0.42"
tqdm = "^4.67.1"
pyyaml = "^6.0.2"
biopython = "^1.85"
psycopg2-binary = "^2.9.10"
pandas = "^2.3.1"
scikit-learn = "^1.7.1"
matplotlib = "^3.10.5"
seaborn = "^0.13.2"
tensorflow = "^2.19.0"
hypothesis = "^6.137.1"
pytest = "^8.4.1"

[tool.poetry.scripts]
app = 'src.main:main'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
